// resources/js/enums/ApplicationStatusEnum.tsx
export enum ApplicationStatusEnum {
    IN_PROCESS = 'in_process',
    FOR_CCD_APPROVAL = 'for_ccd_approval',
    FOR_INSPECTION = 'for_inspection',
    FOR_VERIFICATION = 'for_verification',
    FOR_COLLECTION = 'for_collection',
    FOR_SIGNING = 'for_signing',
    FOR_INSTALLATION_APPROVAL = 'for_installation_approval',
    ACTIVE = 'active',
}